feat(core): add `tree-sitter-breadcrumbs` command
Using this command, the user will be able to see a popup
showing all parent lines for cursor. This is good to get
context on the editor on big files.

This implementation is not ideal, as breadcrumbs should maybe
appear all the time, but is a simple hack to suppress the need
for the time being.